// src/app/models/reclamationClient.ts
export interface ReclamationClient {
  NumeroReclamation?: string;
  DateReclamation?: Date;
  TypeArticle?: string;
  EnChargePar?: string;
  Resolution?: string;
  DateResolution?: Date;
  DecissionDe?: string;
  Execution?: string;
  Observation?: string;
  Description?: string;
  NumeroFiche?: string;
  NumeroVisite?: number;
  CodeArticle?: string;
  DesignationArticle?: string;
}